+++ packages/client/jest.config.js
@@ -8,4 +8,7 @@ export default {
   globals: {
     __SERVER_PORT__: process.env.SERVER_PORT,
   },
+  moduleNameMapper: {
+    "^.+\\.(css|scss)$": "babel-jest"
+  }
 }
