+++ packages/client/src/client.d.ts
@@ -1,8 +1,1 @@
 declare const __SERVER_PORT__: number
-
-declare type Field = {
-  label: string
-  type: string
-  name: string
-  value?: string
-}
