+++ packages/client/src/components/index.ts
@@ -1,0 +1,4 @@
+import Avatar from "./Avatar";
+export {
+    Avatar
+}
