+++ packages/client/src/pages/index.ts
@@ -1,0 +1,4 @@
+import ProfilePage from './profile';
+export {
+    ProfilePage
+}
