+++ packages/client/src/types.d.ts
@@ -1,0 +1,6 @@
+declare type Field = {
+  label: string
+  type: string
+  name: string
+  value?: string
+}
